配置完善， 使可运行和测试。

+++ Canku/src/com/buscode/canku/storage/Shop.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 @GsonBean
 public class Shop {
     public ShopInfo shop;
-    public int totalPrice;
+    public float totalPrice;
     public int totalNum;
     public ShopOrder minLuck;
     public ArrayList<ShopOrder> orders;
